[#35][refactor] user-service와 통신하는 코드 수정
@FeignClient를 사용하여 유저 서비스에서 userId를 확인하기 위해 user-service 코드와 동일하게 변경함.
